// src/api/database/global-types.ts
export interface TestInput {
  index: number;
  input: string;
  output: string;
  isPublic?: boolean | null;
  problemId: any;
}